[Hotfix] Adding support for _id, _type, _index
Previously, esx scroll would not omit any metadata fields about the
original document's ID, document type, or index. These fields are useful
for many reasons, but most obviously they are useful in the scenario of
reindexing data. Therefore, esx scroll was reworked to add _id, _type,
and _index in the emitted documents, and esx index was reworked to look
for these special meta fields in order to propery set the document ID in
the index operation.

Note that currently, the _type and _index fields are unused currently in
esx index, and furthermore, all fields that start with an underscore
("_") are omitted from the document source before indexing.